cleaned featured resources desktop code
