+++ src/components/cards/ResourcesCards/FeaturedResourcesCardDesktop.js
@@ -22,7 +22,7 @@ import {primaryColor,
   infoColor,
   roseColor,
   grayColor,
-  vcColor} from '../../assets/material-kit-assets/jss/material-kit-react'
+  vcColor} from '../../../assets/material-kit-assets/jss/material-kit-react'
 
 const colorMapping = {
   'primary': primaryColor,
@@ -38,8 +38,10 @@ const colorMapping = {
 
 const useStyles = makeStyles({
   root: {
-    height: '386px',
-    align: 'center',
+    height: '145px',
+    width: '95%',
+    display: 'flex',
+    align: 'left',
     position: 'relative',
     boxShadow: "0 2px 2px 0 rgba(0, 0, 0, 0.14), 0 3px 1px -2px rgba(0, 0, 0, 0.2), 0 1px 5px 0 rgba(0, 0, 0, 0.12)",
     transition: 'all 0.3s',
@@ -50,28 +52,28 @@ const useStyles = makeStyles({
   },
   imgOverlay: {
     position:'absolute',
-    background: 'linear-gradient(180deg, rgba(0, 0, 0, 0) 0%, rgba(0, 0, 0, 0.8) 71%)',
+    background: '#FFFFFF',
     left: 0,
     right: 0,
     top: 0,
     bottom: 0,
-    height: '50.26%',
-    opacity:'50%'
+    height: '100%',
+    opacity:'0%'
   },
   media: {
     position: 'absolute',
     left: '0%',
     right: '0%',
     top: '0%',
-    bottom: '50.26%',
-
+    bottom: '100%',
+    paddingRight: "70%",
     borderRadius: '5px 5px 0px 0px'
   },
   title: {
     position: 'absolute',
-    left: '8.28%',
+    left: '35%',
     right: '8.28%',
-    top: '29.02%',
+    top: '25%',
     bottom: '51.04%',
 
     /* Desktop/Body */
@@ -82,15 +84,15 @@ const useStyles = makeStyles({
     fontSize: '25px',
     lineHeight: '30px',
 
-    color: '#FFFFFF'
+    color: '#000000'
   },
   description: {
     // display: 'none',
     // height: 60,
     position: 'absolute',
-    left: '8.28%',
+    left: '35%',
     right: '8.28%',
-    top: '55.44%',
+    top: '50%',
     bottom: '9.59%',
 
     /* Desktop/Details */
@@ -103,6 +105,19 @@ const useStyles = makeStyles({
 
     color: '#000000',
     // overflow: 'hidden'
+  },
+  category: {
+    position: 'absolute',
+    left: '35%',
+    top: '10%',
+
+    fontFamily: 'Poppins',
+    fontStyle: 'normal',
+    fontWeight: 'normal',
+    fontSize: '11px',
+    lineHeight: '16px',
+
+    color: "#636363",
   },
   cardHeader:{
     textTransform: 'capitalize',
@@ -156,7 +171,7 @@ const trimDescription = function(description) {
   return description
 };
 
-export default function FeaturedResourcesCard(props) {
+export default function ResourcesCard(props) {
   const classes = useStyles();
 
   const [anchorEl, setAnchorEl] = React.useState(null);
@@ -167,7 +182,7 @@ export default function FeaturedResourcesCard(props) {
 
 
   let {iosLink, androidLink, website, share, img, title, description, tags,
-    headerTitle, headerColor} = props;
+    headerTitle, headerColor, category} = props;
   if(headerColor && colorMapping.hasOwnProperty(headerColor)){
     headerColor = colorMapping[headerColor.toLowerCase()]
   }
@@ -178,23 +193,24 @@ export default function FeaturedResourcesCard(props) {
   return (
     <Card className={classes.root}>
       <a href={website} target='_blank' rel="noopener noreferrer" style={{color: 'black'}}>
-        <div className={classes.mediaContainer}>
-          {headerTitle && <div className={classes.cardHeader} style={{backgroundColor: headerColor, fontWeight:'bold'}}>{headerTitle}</div>}
-          <CardMedia
-            component="img"
-            height="50.26%"
-            className={classes.media}
-            image={img}
-            title={title}
-          />
-          <div className={classes.imgOverlay}/>
-
-        </div>
+      <div className={classes.mediaContainer}>
+        <CardMedia
+          component="img"
+          height="100%"
+          className={classes.media}
+          image={img}
+          title={title}
+        />
+        <div className={classes.imgOverlay}/>
+      </div>
 
         <CardContent style={{marginBottom: 0}}>
-          <Typography gutterBottom variant="h5" component="h2" className={classes.title} >
-            {title}
-          </Typography>
+        <Typography variant="body2" color="textSecondary" component="p" className={classes.category} >
+        {category}
+        </Typography>
+        <Typography gutterBottom variant="h5" component="h2" className={classes.title} >
+          {title}
+        </Typography>
           <Typography variant="body2" color="textSecondary" component="p" className={classes.description}>
             {trimDescription(description)}
           </Typography>
@@ -205,8 +221,8 @@ export default function FeaturedResourcesCard(props) {
               </Button>
             )
           })}
-
         </CardContent>
+
         <CardActions disableSpacing style={{marginTop: 0, paddingTop: 0, display: 'none',}}>
           <div style={{float: 'left'}}>
             {share &&
@@ -257,7 +273,7 @@ export default function FeaturedResourcesCard(props) {
   );
 }
 
-FeaturedResourcesCard.propTypes = {
+ResourcesCard.propTypes = {
   iosLink: PropTypes.string,
   androidLink: PropTypes.string,
   website: PropTypes.string.isRequired,
@@ -267,5 +283,6 @@ FeaturedResourcesCard.propTypes = {
   title: PropTypes.string.isRequired,
   tags: PropTypes.string.isRequired,
   headerTitle: PropTypes.string,
-  headerColor: PropTypes.string
+  headerColor: PropTypes.string,
+  category: PropTypes.string.isRequired
 };
