+++ src/components/resources/featured/ResourcesFeaturedMobile.js
@@ -1,33 +1,34 @@
 import React from "react";
+import ResourcesListFunctionality from "../resourcesList/ResourcesListFunctionality";
 import {MuiThemeProvider} from "@material-ui/core/styles";
-import Carousel from 'react-material-ui-carousel'
 import Button from "../../material-kit-components/CustomButtons/Button.js";
 
 import {makeStyles} from "@material-ui/core/styles";
 import Data from "../../../assets/ResourcesData";
-import {CustomTheme, CustomButton} from "../..";
+import {CustomTheme, EventCardFeatured, Template} from "../..";
+import ResourcesCardGridView from "../../cards/ResourcesCards/ResourcesCardGridView.js";
 const FeaturedData = Data.FeaturedData;
 const theme = CustomTheme;
-
 
 const containerStyles = makeStyles(() => ({
   container: {
     position: 'relative',
-    background: 'linear-gradient(0deg, rgba(0, 0, 0, 0.4), rgba(0, 0, 0, 0.4) 71%)',
+    display: 'block',
+    background: 'linear-gradient(0deg, rgba(0, 0, 0, 0.4), rgba(0, 0, 0, 0.4))',
     filter: 'blur(15px) contrast(125%) brightness(75%)',
   },
   img: {
     position: 'absolute',
-    left: '7%',
-    top: '8%',
+    left: '0%',
+    top: '7.91%',
     objectFit: 'cover',
   },
   card: {
     position: 'absolute',
-    left: '7%',
-    right: '7%',
-    top: '40.31%',
-    bottom: '14%',
+    left: '3.4%',
+    right: '62.74%',
+    top: '7.91%',
+    bottom: '14.5%',
     background: '#FFFFFF',
   },
   category: {
@@ -35,12 +36,12 @@ const containerStyles = makeStyles(() => ({
     left: '7.55%',
     right: '7.55%',
     top: '7.81%',
-    bottom: '58.82%',
+    bottom: '77.93%',
     fontFamily: 'Poppins',
     fontStyle: 'normal',
     fontWeight: 'normal',
-    fontSize: '12px',
-    lineHeight: '21px',
+    fontSize: '1.2vw',
+    lineHeight: '30px',
     color: '#000000'
   },
   title: {
@@ -48,7 +49,7 @@ const containerStyles = makeStyles(() => ({
     fontFamily: 'Poppins',
     fontStyle: 'normal',
     fontWeight: 'normal',
-    fontSize: 'vw',
+    fontSize: '2vw',
     lineHeight: 'vh',
     color: '#0072CE'
   },
@@ -56,14 +57,13 @@ const containerStyles = makeStyles(() => ({
     background: 'rgba(255, 255, 255, 0.85)',
     position: 'relative',
     marginLeft: '3%',
-    bottom: '42.65%',
     borderRadius: '5px',
     zIndex: 10,
     fontFamily: 'Poppins',
     fontStyle: 'normal',
     fontWeight: 'normal',
-    fontSize: '8px',
-    lineHeight: '6px',
+    fontSize: '10px',
+    lineHeight: '10px',
     textAlign: 'center',
   },
   description: {
@@ -71,71 +71,82 @@ const containerStyles = makeStyles(() => ({
     fontFamily: 'Poppins',
     fontStyle: 'normal',
     fontWeight: 'normal',
-    fontSize: '11px',
-    lineHeight: 'vh',
-    color: '#000000'
+    fontSize: 'vw',
+    lineHeight: '17px',
+    color: '#000000',
+    paddingTop: '5px'
   },
 }));
 
-function Item(props) {
-  const contStyle = containerStyles();
-  return (
-      <div>props</div>
-  )
+class ResourcesFeaturedDesktop extends ResourcesListFunctionality {
+constructor(props){
+  super(props);
+  this.state = {
+    displayResources: []
+  };
 }
+render() {
+  const MAX_RESOURCES_DISPLAYED = 3;
+  let value = 0;
 
-export default function ResourcesFeaturedMobile() {
-  const contStyle = containerStyles();
-  return (
-    <MuiThemeProvider theme={theme}>
-    <Carousel>
-        {
-            Object.keys(FeaturedData).map(key => {
-              let data = FeaturedData[key];
-              return (
-                  <div style={{overflow:'hidden'}}>
-                    <img src={data.img} alt={data.img} height={'440px'} width={'100%'} className={contStyle.container}/>
-                    <img src={data.img} alt={data.img} height={'180px'} width={'86%'} className={contStyle.img}/>
-                    <div className={contStyle.card} >
-                      <p className={contStyle.category}>
-                        {data.category.category
-                          .split(' ')
-                          .map(function(word) {
-                              return word[0].toUpperCase() + word.substr(1);
-                          })
-                          .join(' ')
-                        }
+    return (
+      <MuiThemeProvider theme={theme}>
+      <div style={{backgroundColor: "#3B5998",
+          height: "600px",
+          width: "104.2%",
+          borderStyle: "solid",
+          borderColor: "#3B5998",
+          borderWidth: "thick",
+          flexDirection: "row",
+          display: "flex",
+          marginLeft: "-4%"}}
+          >
+          <div style={{marginTop: "20px", marginLeft: "6%", color:"white", textAlign: "left"}}>
 
-                        <p className={contStyle.title}>
-                        {data.title}
-                        </p>
+          </div>
+          <div style={{backgroundColor: "#FB750D",
+            marginTop: "-5px",
+            height: "550px",
+            marginLeft: "80%",
+            width: "14.6%",
+            marginRight: "-0.25%",
+            borderColor: "#FB750D",
+            borderRadius: "0px 0px 0px 102px"
+          }}>
+            <div style={{flexDirection: "column", display: "flex", marginTop: "25px", marginLeft: "-540%",}}>
+              <h2 style={{fontSize:40, color:"white"}}>Featured Resources</h2>
+              <p style={{fontSize: 20, color:"white"}}>Some of our most popular resources to help you tackle recent events, be mindful of your health, improve your job search, and more.
+                <br /><br />
+              </p>
+            </div>
+            <div style={{flexDirection: "row", display: "flex", marginLeft: "-540%",}}>
+              { Object.keys(FeaturedData).map(key => {
+                  if(value < MAX_RESOURCES_DISPLAYED){
+                    let data = FeaturedData[key];
+                    value = value + 1;
+                    return(
+                      <ResourcesCardGridView
+                      website={data.links.website}
+                      img={data.img}
+                      title={data.title}
+                      description={data.description}
+                      iosLink={data.links.iosLink}
+                      androidLink={data.links.androidLink}
+                      tags={data.category.tags}
+                      category={data.category.category}
+                      share/>
+                    );
+                  }
+              })}
+            </div>
+          </div>
+      </div>
 
-                        <p>
-                          {data.category.tags.map(ele => {
-                            return (
-                              <CustomButton text={ele} size={'small'} color={"blue"} disabled className={contStyle.button}/>
-                            )
-                          })}
-                        </p>
 
-                        <p className={contStyle.description} style={{marginTop: 10, position:'relative'}}>
-                          {data.description}
-                        </p>
 
-                        <p>
-                          <CustomButton text={"Explore"} href={data.links.website}
-                            color={"orange"} size={"medium"} style={{marginTop: 3, position:'relative'}}/>
-                        </p>
+      </MuiThemeProvider>
+    )
+  }
 
-                      </p>
-
-                    </div>
-                  </div>
-              )
-
-            })
-        }
-    </Carousel>
-    </MuiThemeProvider>
-  )
 }
+export default ResourcesFeaturedDesktop;
